Migrate all code to Spark MAX
All code has been changed to use Spark MAXes which now are only 1 motor
on each side of the bot which results in less parameters for the drive
methods.

+++ src/main/java/org/usfirst/frc/team2473/robot/commands/AutonomousTester.java
@@ -25,7 +25,7 @@ public class AutonomousTester extends CommandGroup {
 	 * @param turnAngle 		the angle to turn (degrees)
 	 * @param distanceSecond	the distance to drive after turning (inches)
 	 */
-	public void addDriveTurnDrive(int distanceFirst, int turnAngle, int distanceSecond) {
+	public void addDriveTurnDrive(double distanceFirst, int turnAngle, double distanceSecond) {
 		addSequential(new StraightDrive(distanceFirst, 0.3));
 		addSequential(new WaitCommand(WAIT));
 		addSequential(new PointTurn(turnAngle, 0.45));
